+++ src/components/Links/index.tsx
@@ -1,32 +1,20 @@
 import styles from './Links.module.css'
 
-const ITEMS = [
-  {
-    href: "https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app",
-    title: "Docs",
-    description: "Find in-depth information about Next.js features and API."
-  },
-  {
-    href: "https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app",
-    title: "Learn",
-    description: "Learn about Next.js in an interactive course with quizzes!"
-  },
-  {
-    href: "https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app",
-    title: "Templates",
-    description: "Discover and deploy boilerplate example Next.js projects."
-  },
-  {
-    href: "https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app",
-    title: "Deploy",
-    description: "Instantly deploy your Next.js site to a shareable URL with Vercel."
-  }
-]
+interface Item {
+  href: string;
+  title: string;
+  description: string;
+}
 
-export function Links() {
+interface LinksProps {
+  items: Item[];
+}
+
+export function Links({ items }: LinksProps) {
+
   return (
     <div className={styles.grid}>
-      {ITEMS.map((item) => {
+      {items.map((item: Item) => {
         return(
           <a
             key={item.href}
